==> r_skripts/0p0078125_scen/00_spust_prikazy.r <==
source("R/01_nacti_default_data.r")
source("R/02_nacti_lai_data.r")
source("R/03_nacti_pozorovani.r")
source("R/04_orez_simulace.r")
source("R/05_vypocty_prumeru.r")
source("R/06_statistiky_rocni_objem_mdni.r")
source("R/07_dl_prumery_srovnani.r")
source("R/08_cary_prekroceni.r")
source("R/09_mapy_zmeny_prutoku.r")
source("R/10_mapy_zmeny_mdenni.r")
source("R/11_mesicni_srovnani_obdobi.r")
source("R/12_violin_denni_scenare.r")
source("R/13_boxplot_denni_obdobi.r")
source("R/14_dl_prumery_lai.r")
source("R/15_cary_prekroceni_lai.r")
source("R/16_mapy_zmeny_lai.r")
